Animations, Double Camera Views , and Scripting
Adding Rifle Run animation and adding appropriate scripts to support running, as well as adding blocks to certain actions depending on the current value of the properties in Animator, for example blocking the ability to run while aiming. Adding multiple camera views  depending if the character is aiming or not.